 \name{TRdata}
     \docType{data}
     \alias{Gauss2D}
     \alias{Gauss2D.learn}
     \alias{Gauss2D.test}
     \alias{Gauss2D.roc}
     \alias{Gauss2DEasy}
     \alias{Gauss2DEasy.learn}
     \alias{Gauss2DEasy.test}
     \alias{Gauss2DEasy.roc}

\alias{Gauss20DFar}
     \alias{Gauss20DFar.learn}
     \alias{Gauss20DFar.test}
     \alias{Gauss20DFar.roc}
	\alias{Gauss20DClose}
     \alias{Gauss20DClose.learn}
     \alias{Gauss20DClose.test}
     \alias{Gauss20DClose.roc}
 \alias{Pyr2D}
     \alias{Pyr2D.learn}
     \alias{Pyr2D.test}
     \alias{Pyr2D.roc}


     \title{Artificial data sets for the TreeRank Package}
     \description{
	Artificial data sets to illustrate the \code{\link{TreeRank}}
        procedure.

       
        For each data set, the \code{.learn} dataframe denotes the learning set, the \code{.test} denotes
the test set and the \code{.roc} denotes the target ROC curve (i.e. the optimal one).

	\code{Gauss2D}, \code{Gauss20DFar} and \code{Gauss20DClose} are generated by a mixture of gaussian in 2 dimensions and in 20 dimensions.

	\code{Pyr2D} is generated by a mixture of uniform distributions in 2 dimensions.
     }

     \format{Dataframes with 3 or 21 dimensions, the response variable is always \code{class} and the positive value \code{1}.}
     \source{Artificial generation.}
     \keyword{datasets}

